start moving and adapting tests
